Removing testing code from production
and more polishing
