+++ peer.go
@@ -5,7 +5,6 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -337,10 +336,7 @@ func (peer *Peer) OnCTRLMsg(msg webrtc.DataChannelMessage) {
 		var authArgs AuthArgs
 		err = json.Unmarshal(raw, &authArgs)
 		// if we're running a test. authorzied only the test token
-		// next line copied from: https://stackoverflow.com/a/45913089/66595
-		if strings.HasSuffix(os.Args[0], ".test") &&
-			authArgs.Token == AValidTokenForTests ||
-			IsAuthorized(authArgs.Token) {
+		if IsAuthorized(authArgs.Token) {
 			peer.Authenticated = true
 			err = json.Unmarshal(raw, &authArgs)
 			peer.Token = authArgs.Token
