+++ test_tokens
@@ -1,0 +1,1 @@
+THEoneANDonlyTOKEN
